+++ src/server/decode.rs
@@ -134,7 +134,7 @@ where
         let reader = Arc::new(Mutex::new(reader.take(len)));
         req.set_body(Body::from_reader(
             BufReader::new(ReadNotifier::new(reader.clone(), body_read_sender)),
-            Some(len as usize),
+            Some(len as u64),
         ));
         Ok(Some((req, BodyReader::Fixed(reader))))
     } else {
